[feat] Group 화면 재구성 및 정리
- Group List function을 제거하고  GroupListScreen 함수 내부에 구현
- connect GroupListScreen => connect GroupScreen
- GroupRegisterScreen을 추가하고 Stack navigator에 screen stack 추가
