+++ src/screens/GroupScreen.js
@@ -1,10 +1,14 @@
 import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
 import { connect } from 'react-redux';
+import { FlatList, View } from 'react-native';
 
 import styled from 'styled-components';
-import GroupList from './GroupList';
-import AddGroupLayout from './AddGroupLayout';
+import GroupRegisterScreen from './GroupRegisterScreen';
+import { FloatingButton } from '~/components/FloatingButton';
+import GroupListEntry from '~/components/GroupListEntry';
+
+import image from '~/utils/image_sample';
 
 const Container = styled.SafeAreaView`
   flex: 1;
@@ -12,7 +16,8 @@ const Container = styled.SafeAreaView`
 
 const Stack = createStackNavigator();
 
-function GroupScreen() {
+function GroupScreen(props) {
+  let getGroup = props.getGroup;
   fetch(
     'http://ec2-13-125-12-178.ap-northeast-2.compute.amazonaws.com:8080/users/1/groups',
     {
@@ -38,22 +43,55 @@ function GroupScreen() {
     <Stack.Navigator>
       <Stack.Screen
         name="Group List"
-        component={connect(
-          mapStateToProps,
-          mapDispatchToProps
-        )(GroupListScreen)}
+        component={GroupListScreen}
+      />
+      <Stack.Screen
+        name="Group Register"
+        component={GroupRegisterScreen}
       />
     </Stack.Navigator>
   );
 }
 
-function GroupListScreen(props) {
-  let getGroup = props.getGroup;
+const SampleList = [
+  {
+    party: '그룹 1',
+    date: '2020-08-08',
+    location: 'Seongnam',
+    gameTitle: 'Zenga',
+    members: [image[0], image[1], image[2], image[3]],
+  },
+  {
+    party: '그룹 2',
+    date: '2020-08-28',
+    location: 'Seould',
+    gameTitle: 'Gizmo',
+    members: [image[3], image[4], image[5], image[6]],
+  },
+];
 
+function GroupListScreen(props) {
+  let data = props.groupList;
+  if (!data) data = SampleList;
   return (
     <Container>
-      <GroupList />
-      <AddGroupLayout />
+      <View>
+        <FlatList
+          data={data}
+          renderItem={({ item }) => (
+            <GroupListEntry
+              item={{
+                members: item.members,
+                gameTitle: item.image,
+                date: item.date,
+                party: item.party,
+                location: item.location,
+              }}
+            />
+          )}
+        />
+        <FloatingButton onPress={()=> { props.navigation.navigate('Group Register'); }} />
+      </View>
     </Container>
   );
 }
@@ -71,4 +109,4 @@ function mapDispatchToProps(dispatch) {
   };
 }
 
-export default GroupScreen;
+export default connect(mapStateToProps, mapDispatchToProps)(GroupScreen);
